+++ src/sketches/023_fdn/graph/patches/patches.hpp
